renamed projectile class back to bullet for easier review
